:green_heart: fix the problem with the import. mysteriously, moban.extensions works but relocating it to moban.jinja2.extension wont do
