+++ docs/level-7-use-custom-jinja2-filter-test-n-global/custom-jj2-plugin/filter.py
@@ -1,7 +1,7 @@
 import sys
 import base64
 
-from moban.jinjia2.extensions import JinjaFilter
+from moban.extensions import JinjaFilter
 
 
 @JinjaFilter()
