+++ docs/level-7-use-custom-jinja2-filter-test-n-global/custom-jj2-plugin/test.py
@@ -1,4 +1,4 @@
-from moban.jinja2.extensions import JinjaTest
+from moban.extensions import JinjaTest
 
 
 @JinjaTest()
