+++ moban/jinja2/filters/text.py
@@ -1,6 +1,6 @@
 import re
 
-from moban.jinja2.extensions import JinjaFilter
+from moban.extensions import JinjaFilter
 
 
 @JinjaFilter()
