+++ tests/test_engine.py
@@ -6,10 +6,10 @@ import moban.exceptions as exceptions
 from mock import patch
 from nose.tools import eq_, raises
 from moban.plugins import ENGINES, expand_template_directories
+from moban.extensions import jinja_global
 from moban.jinja2.engine import Engine
 from moban.engine_factory import Context
 from moban.engine_handlebars import EngineHandlebars
-from moban.jinja2.extensions import jinja_global
 
 
 @PluginInfo("library", tags=["testmobans"])
